Finish restructuring source tree.
Fix template paths.

Fix import paths.
