+++ c_test_environment/clang_datalog_tests.py
@@ -5,7 +5,7 @@ from generate_test_relations import generate_default
 from generate_test_relations import need_generate
 import raco.backends.clang as clang
 import raco.backends.clang.clangcommon as clangcommon
-from platform_tests import DatalogPlatformTest
+from raco.platform_tests import DatalogPlatformTest
 
 import sys
 sys.path.append('./examples')
