+++ c_test_environment/clang_myrial_tests.py
@@ -4,7 +4,7 @@ from testquery import ClangRunner
 from generate_test_relations import generate_default
 from generate_test_relations import need_generate
 from raco.backends.clang import CCAlgebra
-from platform_tests import MyriaLPlatformTestHarness, MyriaLPlatformTests
+from raco.platform_tests import MyriaLPlatformTestHarness, MyriaLPlatformTests
 from raco.compile import compile
 
 import sys
