+++ raco/datalog/datalog_test.py
@@ -3,7 +3,7 @@ import json
 
 import raco.fakedb
 from raco import RACompiler
-from raco.language.myrialang import (compile_to_json,
+from raco.backends.myria import (compile_to_json,
                                      MyriaLeftDeepTreeAlgebra,
                                      MyriaHyperCubeAlgebra)
 from raco.catalog import FakeCatalog
