+++ raco/operator_test.py
@@ -97,7 +97,7 @@ class OperatorTest(unittest.TestCase):
         self.assertEqual([x[0] for x in result][-1], 37857)
 
         # test whether we can generate json without errors
-        from raco.language.myrialang import (compile_to_json,
+        from raco.backends.myria import (compile_to_json,
                                              MyriaLeftDeepTreeAlgebra)
         from compile import optimize
         import json
